more work on project IO
git-svn-id: http://cars9.uchicago.edu/svn/demeter/trunk@462 a48c6eca-4e1b-0410-a7b3-d802d4c683d2

+++ lib/Demeter/Fit.pm
@@ -1133,13 +1133,14 @@ override 'serialize' => sub {
   close $PLOT;
 
 
-
-  my $readme = File::Spec->catfile($self->share_folder, "Readme.fit_serialization");
-  my $target = File::Spec->catfile($self->folder, "Readme");
-  copy($readme, $target);
-  if (not $args{nozip}) {
-    $self->zip_project($self->folder, $args{file});
-    rmtree($self->folder);
+  if ($args{file}) {
+    my $readme = File::Spec->catfile($self->share_folder, "Readme.fit_serialization");
+    my $target = File::Spec->catfile($self->folder, "Readme");
+    copy($readme, $target);
+    if (not $args{nozip}) {
+      $self->zip_project($self->folder, $args{file});
+      rmtree($self->folder);
+    };
   };
 
   return $self;
